+++ client/packages/cli/src/cli.ts
@@ -4,8 +4,8 @@ import { Command } from "commander"
 import { handleInitCmd, initConfigFile } from "./commands/init.ts"
 import { CONFIG_FILE } from "@/consts.ts"
 import { wrapCryptoWaitReady } from "./utils/fns.ts"
-import { handleRegisterCmd } from "./commands/register/register.ts"
-import { handleRebootCommand } from "./commands/reboot/index.ts"
+import { handleRegisterCmd } from "./commands/registerGateway/register.ts"
+import { handleRebootCommand } from "./commands/rebootGateway/index.ts"
 import { handleSubmitCmd } from "./commands/submit/submit.ts"
 import { handleBidCmd } from "./commands/bid.ts"
 import { handleDgfCmd } from "./commands/dgf.ts"
