+++ client/packages/cli/src/commands/rebootGateway/index.ts
