+++ client/packages/cli/src/commands/registerGateway/__test__/gateway.test.ts
