+++ client/packages/cli/src/commands/registerGateway/gateway.ts
