+++ client/packages/cli/src/commands/registerGateway/vendors/eth.ts
