+++ client/packages/cli/src/commands/registerGateway/vendors/substrate.ts
